requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
black==24.3.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.11
django-git-info==0.1.1
django-jinja==2.11.0
django-ninja==1.1.0
django-picklefield==3.2
django-q2==1.6.2
django-role-permissions==3.2.0
flake8==7.0.0
flake8-print==5.0.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pycodestyle==2.11.1
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
PyJWT==2.8.0
requests==2.32.3
six==1.16.0
smmap==5.0.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
